Moved country code check from standard to rule, added some better error messages to format rules
