+++ lib/rules/core/fields-match-standard-rule-spec.js
@@ -16,62 +16,6 @@ describe('FieldsMatchStandardRule', function() {
         expect(isTargeted).toBe(true);
     });
 
-    // ISO3166-1-ALPHA2
-    it('should return no error for an valid country code', function() {
-        let model = new Model({
-            "type": "Event",
-            "fields": {
-                "country": {
-                    "fieldName": "country",
-                    "requiredType": "http://schema.org/Text",
-                    "standard": "$ISO3166-1-ALPHA2"
-                }
-            }
-        });
-        let values = [
-            "GB",
-            "CN",
-            "US"
-        ];
-
-        for (let value of values) {
-            let data = {
-                "country": value
-            };
-            let errors = rule.validate(data, model, null);
-            expect(errors.length).toBe(0);
-        }
-    });
-    it('should return an error for an invalid country code', function() {
-        let model = new Model({
-            "type": "Event",
-            "fields": {
-                "country": {
-                    "fieldName": "country",
-                    "requiredType": "http://schema.org/Text",
-                    "standard": "$ISO3166-1-ALPHA2"
-                }
-            }
-        });
-
-        let values = [
-            "BC",
-            "QB",
-            "ZY",
-            "A"
-        ];
-
-        for (let value of values) {
-            let data = {
-                "country": value
-            };
-            let errors = rule.validate(data, model, null);
-            expect(errors.length).toBe(1);
-            expect(errors[0].type).toBe(ValidationErrorType.INVALID_STANDARD);
-            expect(errors[0].severity).toBe(ValidationErrorSeverity.FAILURE);
-        }
-    });
-
     // WGS84Latitude
     it('should return no error for an valid latitude', function() {
         let model = new Model({
