+++ lib/rules/format/country-code-format-rule.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const cc = require('currency-codes');
+const validator = require('validator');
 const Rule = require('../rule');
 const Field = require('../../classes/field');
 const ValidationError = require('../../errors/validation-error');
@@ -8,12 +8,12 @@ const ValidationErrorType = require('../../errors/validation-error-type');
 const ValidationErrorCategory = require('../../errors/validation-error-category');
 const ValidationErrorSeverity = require('../../errors/validation-error-severity');
 
-module.exports = class CurrencyCodeFormatRule extends Rule {
+module.exports = class CountryCodeFormatRule extends Rule {
 
     constructor(options) {
         super(options);
         this._targetFields = '*';
-        this._description = "Validates that currency code fields are in the correct format.";
+        this._description = "Validates that country code fields are in the correct format.";
     }
 
     validateField(data, field, model, parent) {
@@ -22,14 +22,14 @@ module.exports = class CurrencyCodeFormatRule extends Rule {
         }
         let errors = [];
         let fieldObj = new Field(model.fields[field]);
-        if (fieldObj.sameAs === 'http://schema.org/priceCurrency') {
-            if (typeof(cc.code(data[field])) === 'undefined') {
+        if (fieldObj.sameAs === 'http://schema.org/addressCountry') {
+            if (!validator.isISO31661Alpha2(data[field])) {
                 errors.push(
                     new ValidationError(
                         {
                             "category": ValidationErrorCategory.CONFORMANCE,
                             "type": ValidationErrorType.INVALID_FORMAT,
-                            "message": "Currency codes should be expressed as per the assigned 3-letter codes in ISO 4217",
+                            "message": "Country codes should be expressed as per the assigned codes in ISO 3166-1 Alpha 2",
                             "value": data[field],
                             "severity": ValidationErrorSeverity.FAILURE,
                             "path": field
