+++ lib/rules/format/date-format-rule.js
@@ -35,6 +35,7 @@ module.exports = class DateFormatRule extends Rule {
                         {
                             "category": ValidationErrorCategory.CONFORMANCE,
                             "type": ValidationErrorType.INVALID_FORMAT,
+                            "message": "Dates should be expressed as ISO 8601 dates",
                             "value": data[field],
                             "severity": ValidationErrorSeverity.FAILURE,
                             "path": field
