+++ lib/rules/format/datetime-format-rule.js
@@ -32,6 +32,7 @@ module.exports = class DatetimeFormatRule extends Rule {
                         {
                             "category": ValidationErrorCategory.CONFORMANCE,
                             "type": ValidationErrorType.INVALID_FORMAT,
+                            "message": "DateTimes should be expressed as ISO 8601 format datetimes with a trailing definition of timezone",
                             "value": data[field],
                             "severity": ValidationErrorSeverity.FAILURE,
                             "path": field
