+++ lib/rules/format/duration-format-rule.js
@@ -31,6 +31,7 @@ module.exports = class DurationFormatRule extends Rule {
                         {
                             "category": ValidationErrorCategory.CONFORMANCE,
                             "type": ValidationErrorType.INVALID_FORMAT,
+                            "message": "Durations should be expressed as ISO 8601 durations",
                             "value": data[field],
                             "severity": ValidationErrorSeverity.FAILURE,
                             "path": field
