+++ package.json
@@ -25,7 +25,8 @@
   "dependencies": {
     "currency-codes": "^1.3.0",
     "jasmine-node": "^1.15.0",
-    "moment": "^2.22.2"
+    "moment": "^2.22.2",
+    "validator": "^10.4.0"
   },
   "scripts": {
     "test": "./node_modules/.bin/jasmine-node lib"
